+++ ts-client/src/alpha-vault/index.ts
@@ -17,6 +17,7 @@ import {
 import {
   ALPHA_VAULT_TREASURY_ID,
   DYNAMIC_AMM_PROGRAM_ID,
+  MERKLE_PROOF_API,
   PROGRAM_ID,
   VaultPoint,
   VaultState,
@@ -35,6 +36,7 @@ import {
   createDlmmProgram,
   createDammProgram,
   createCpAmmProgram,
+  deriveMerkleProofMetadata,
 } from "./helper";
 import IDL from "./alpha_vault.json";
 import {
@@ -267,10 +269,8 @@ export class AlphaVault {
 
     if (vault.poolType === PoolType.DAMMV2) {
       const cpAmm = createCpAmmProgram(connection, opt);
-      const pool = (await cpAmm.account.pool.fetch(
-        vault.pool
-      ))
-      
+      const pool = await cpAmm.account.pool.fetch(vault.pool);
+
       return new AlphaVault(
         program,
         vaultAddress,
@@ -549,19 +549,20 @@ export class AlphaVault {
       .mul(escrowAccount.totalDeposit)
       .div(this.vault.totalDeposit);
     const totalClaimed = escrowAccount.claimedToken;
+
     const totalClaimable = (() => {
-      const currentSlotBN = new BN(
+      const currentPoint = new BN(
         this.vault.activationType === ActivationType.SLOT
           ? currentSlot
           : currentTimestamp
       );
-      if (currentSlotBN.lt(this.vault.startVestingPoint)) {
+      if (currentPoint.lt(this.vault.startVestingPoint)) {
         return new BN(0);
       }
 
-      const endSlot = BN.min(currentSlotBN, this.vault.endVestingPoint);
+      const endPoint = BN.min(currentPoint, this.vault.endVestingPoint);
       const totalClaimableToken = this.vault.boughtToken
-        .mul(endSlot.add(new BN(1)).sub(this.vault.startVestingPoint))
+        .mul(endPoint.add(new BN(1)).sub(this.vault.startVestingPoint))
         .div(
           this.vault.endVestingPoint
             .add(new BN(1))
@@ -570,6 +571,7 @@ export class AlphaVault {
       const drippedEscrowAmount = totalClaimableToken
         .mul(escrowAccount.totalDeposit)
         .div(this.vault.totalDeposit);
+
       return drippedEscrowAmount.sub(escrowAccount.claimedToken);
     })();
 
@@ -632,8 +634,12 @@ export class AlphaVault {
     const canDeposit = this.canDeposit(escrow, merkleProof);
     const hadDeposited = escrow ? escrow.totalDeposit.gtn(0) : false;
     const canWithdraw = this.canWithdraw(escrow);
+    const canWithdrawDepositOverflow = this.canWithdrawDepositOverflow(escrow);
+    const availableDepositOverflow =
+      this.escrowAvailableDepositOverflowAmount(escrow);
     const canWithdrawRemainingQuote = this.canWithdrawRemainingQuote(escrow);
     const hadWithdrawnRemainingQuote = this.hadWithdrawnRemainingQuote(escrow);
+
     return {
       claimInfo,
       depositInfo,
@@ -644,6 +650,8 @@ export class AlphaVault {
       canDeposit,
       hadDeposited,
       canWithdraw,
+      canWithdrawDepositOverflow,
+      availableDepositOverflow,
       canWithdrawRemainingQuote,
       hadWithdrawnRemainingQuote,
     };
@@ -716,6 +724,45 @@ export class AlphaVault {
     return false;
   }
 
+  private canWithdrawDepositOverflow(escrow: Escrow | null) {
+    if (!escrow) return false;
+
+    const currentPoint =
+      this.vault.activationType === ActivationType.SLOT
+        ? this.clock.slot.toNumber()
+        : this.clock.unixTimestamp.toNumber();
+
+    const escrowAvailableDepositOverflowAmount =
+      this.escrowAvailableDepositOverflowAmount(escrow);
+
+    return (
+      currentPoint > this.vaultPoint.lastJoinPoint &&
+      currentPoint <= this.vaultPoint.lastBuyingPoint &&
+      escrowAvailableDepositOverflowAmount.gt(new BN(0))
+    );
+  }
+
+  private escrowAvailableDepositOverflowAmount(escrow: Escrow | null) {
+    if (
+      !escrow ||
+      this.vault.vaultMode == VaultMode.FCFS ||
+      this.vault.totalDeposit.isZero()
+    )
+      return new BN(0);
+
+    const vaultDepositOverflow = this.vault.totalDeposit.gt(
+      this.vault.maxBuyingCap
+    )
+      ? this.vault.totalDeposit.sub(this.vault.maxBuyingCap)
+      : new BN(0);
+
+    const escrowDepositOverflow = vaultDepositOverflow
+      .mul(escrow.totalDeposit)
+      .div(this.vault.totalDeposit);
+
+    return escrowDepositOverflow.sub(escrow.withdrawnDepositOverflow);
+  }
+
   private canWithdrawRemainingQuote(escrow: Escrow | null) {
     if (!escrow) return false;
 
@@ -724,15 +771,15 @@ export class AlphaVault {
         ? this.clock.slot.toNumber()
         : this.clock.unixTimestamp.toNumber();
 
-    // make sure the user can withdraw after alpha-vault has done purchasing process
-    if (
-      this.vault.totalDeposit.gt(this.vault.swappedAmount) &&
-      currentPoint > this.vaultPoint.lastBuyingPoint
-    ) {
-      return true;
-    }
+    const remainingQuoteAmount = this.vault.totalDeposit.sub(
+      this.vault.swappedAmount
+    );
 
-    return false;
+    return (
+      currentPoint > this.vaultPoint.lastBuyingPoint &&
+      remainingQuoteAmount.gt(new BN(0)) &&
+      escrow.refunded === 0
+    );
   }
 
   private hadWithdrawnRemainingQuote(escrow: Escrow | null) {
@@ -1178,6 +1225,135 @@ export class AlphaVault {
         systemProgram: SystemProgram.programId,
       })
       .transaction();
+  }
+
+  /**
+   * Creates a Merkle proof metadata for the vault.
+   *
+   * @param {PublicKey} vaultCreator - The public key of the creator's wallet.
+   * @param {string} proofUrl - The URL pointing to the Merkle proof data.
+   * @return {Promise<Transaction>} A promise that resolves to the transaction for creating the Merkle proof metadata.
+   */
+
+  public async createMerkleProofMetadata(
+    vaultCreator: PublicKey,
+    proofUrl: string
+  ) {
+    const [merkleProofMetadata] = deriveMerkleProofMetadata(
+      this.pubkey,
+      this.program.programId
+    );
+
+    const { blockhash, lastValidBlockHeight } =
+      await this.program.provider.connection.getLatestBlockhash("confirmed");
+
+    const createMerkleProofMetadataTx = await this.program.methods
+      .createMerkleProofMetadata(proofUrl)
+      .accountsPartial({
+        merkleProofMetadata,
+        vault: this.pubkey,
+        admin: vaultCreator,
+        systemProgram: SystemProgram.programId,
+      })
+      .transaction();
+
+    return new Transaction({
+      blockhash,
+      lastValidBlockHeight,
+      feePayer: vaultCreator,
+    }).add(createMerkleProofMetadataTx);
+  }
+
+  /**
+   * Retrieves the URL of the Merkle proof data associated with the vault.
+   *
+   * @return {Promise<string>} A promise that resolves to the URL of the Merkle proof data.
+   */
+
+  public async getMerkleProofUrl(): Promise<string> {
+    const [merkleProofMetadata] = deriveMerkleProofMetadata(
+      this.pubkey,
+      this.program.programId
+    );
+
+    const merkleProofMetadataState =
+      await this.program.account.merkleProofMetadata.fetchNullable(
+        merkleProofMetadata
+      );
+
+    return merkleProofMetadataState
+      ? merkleProofMetadataState.proofUrl
+      : MERKLE_PROOF_API[this.opt.cluster || "mainnet-beta"];
+  }
+
+  /**
+   * Retrieves the Merkle proof data required for depositing into the vault.
+   * The data is fetched from the URL stored in the vault's Merkle proof metadata.
+   *
+   * @param {PublicKey} owner - The public key of the depositor's wallet.
+   * @return {Promise<DepositWithProofParams | null>} A promise that resolves to the Merkle proof data if it exists, or null otherwise.
+   */
+  public async getMerkleProofForDeposit(
+    owner: PublicKey
+  ): Promise<DepositWithProofParams | null> {
+    interface MerkleProofResponse {
+      merkle_root_config: String;
+      max_cap: number;
+      proof: number[][];
+    }
+
+    let baseUrl = await this.getMerkleProofUrl();
+    if (baseUrl.endsWith("/")) {
+      baseUrl = baseUrl.slice(0, baseUrl.length - 1);
+    }
+    const fullUrl = `${baseUrl}/${this.pubkey.toBase58()}/${owner.toBase58()}`;
+
+    try {
+      const response = await fetch(fullUrl);
+      if (response.ok) {
+        const data: MerkleProofResponse = await response.json();
+        return {
+          merkleRootConfig: new PublicKey(data.merkle_root_config),
+          maxCap: new BN(data.max_cap),
+          proof: data.proof,
+        };
+      }
+    } catch (e) {
+      console.error(e);
+    }
+    return null;
+  }
+
+  /**
+   * Closes the Merkle proof metadata for the vault.
+   *
+   * @param {PublicKey} vaultCreator - The public key of the creator's wallet.
+   * @return {Promise<Transaction>} A promise that resolves to the transaction for closing the Merkle proof metadata.
+   */
+  public async closeMerkleProofMetadata(vaultCreator: PublicKey) {
+    const [merkleProofMetadata] = deriveMerkleProofMetadata(
+      this.pubkey,
+      this.program.programId
+    );
+
+    const closeMerkleProofMetadataTx = await this.program.methods
+      .closeMerkleProofMetadata()
+      .accountsPartial({
+        merkleProofMetadata,
+        vault: this.pubkey,
+        rentReceiver: vaultCreator,
+        admin: vaultCreator,
+      })
+      .transaction();
+
+    const { blockhash, lastValidBlockHeight } =
+      await this.program.provider.connection.getLatestBlockhash("confirmed");
+
+    return new Transaction({
+      blockhash,
+      lastValidBlockHeight,
+      feePayer: vaultCreator,
+    }).add(closeMerkleProofMetadataTx);
   }
 
   /**